refactor(studio): extract inserters + ensureTables + syncRegionRegistryAndConfig from jobs.ts
