+++ .cortex/skills/build-routing-solution/openrouteservice_app/services/ors_control_app/server/studio/ensure-tables.ts
@@ -1,0 +1,98 @@
+// Studio table DDL ensure helper. Idempotent CREATE TABLE IF NOT EXISTS for
+// the FACT/DIM tables that startGeneration() writes into. Extracted from
+// jobs.ts so the orchestrator stays focused on job lifecycle.
+
+import { log } from '../diagnostics.js';
+import { UNIFIED_DB, UNIFIED_SCHEMA } from './sql-helpers.js';
+
+type SnowSqlFn = (sql: string, database?: string, schema?: string) => Promise<any[]>;
+
+export async function ensureTables(snowSql: SnowSqlFn): Promise<void> {
+  const ddls: { sql: string; db: string; schema: string }[] = [
+    { sql: `CREATE TABLE IF NOT EXISTS ${UNIFIED_DB}.${UNIFIED_SCHEMA}.FACT_VEHICLE_TELEMETRY (
+      TELEMETRY_ID VARCHAR, REGION VARCHAR(100), VEHICLE_TYPE VARCHAR(20),
+      VEHICLE_ID VARCHAR, TRIP_ID VARCHAR,
+      TS TIMESTAMP_NTZ, LATITUDE FLOAT, LONGITUDE FLOAT, POINT_GEOM GEOGRAPHY,
+      SPEED_KMH FLOAT, HEADING_DEG FLOAT, POSTED_SPEED_KMH FLOAT,
+      STATUS VARCHAR(30), IS_SPEEDING BOOLEAN, IS_HOS_VIOLATION BOOLEAN, IS_DETOUR BOOLEAN,
+      GPS_ACCURACY_M FLOAT, LOCATION_ID VARCHAR, LOCATION_TYPE VARCHAR(30),
+      ORS_PROFILE VARCHAR(30), BATTERY_PCT FLOAT, ODOMETER_KM FLOAT, POINT_INDEX INT,
+      JOB_ID VARCHAR
+    ) COMMENT = '{"origin":"sf_sit-is-fleet","name":"oss-build-routing-solution","version":{"major":1,"minor":0},"attributes":{"is_quickstart":1,"source":"sql"}}'`, db: UNIFIED_DB, schema: UNIFIED_SCHEMA },
+    { sql: `CREATE TABLE IF NOT EXISTS ${UNIFIED_DB}.${UNIFIED_SCHEMA}.FACT_TRIPS (
+      TRIP_ID VARCHAR, VEHICLE_ID VARCHAR, DRIVER_ID VARCHAR,
+      VEHICLE_TYPE VARCHAR(20), REGION VARCHAR(100),
+      ORIGIN_POI_ID VARCHAR, DESTINATION_POI_ID VARCHAR,
+      ORIGIN_LAT FLOAT, ORIGIN_LON FLOAT, ORIGIN GEOGRAPHY,
+      DESTINATION_LAT FLOAT, DESTINATION_LON FLOAT, DESTINATION GEOGRAPHY,
+      ROUTE_GEOG GEOGRAPHY, DISTANCE_KM FLOAT, DURATION_MINUTES FLOAT,
+      PLANNED_ROUTE_GEOG GEOGRAPHY, PLANNED_DISTANCE_KM FLOAT,
+      IS_DETOUR BOOLEAN, DETOUR_DISTANCE_KM FLOAT,
+      TRIP_START TIMESTAMP_NTZ, TRIP_END TIMESTAMP_NTZ,
+      STATUS VARCHAR(20), ORS_PROFILE VARCHAR(30),
+      JOB_ID VARCHAR
+    ) COMMENT = '{"origin":"sf_sit-is-fleet","name":"oss-build-routing-solution","version":{"major":1,"minor":0},"attributes":{"is_quickstart":1,"source":"sql"}}'`, db: UNIFIED_DB, schema: UNIFIED_SCHEMA },
+    { sql: `CREATE TABLE IF NOT EXISTS ${UNIFIED_DB}.${UNIFIED_SCHEMA}.DIM_FLEET (
+      VEHICLE_ID VARCHAR, REGION VARCHAR(100), VEHICLE_TYPE VARCHAR(20),
+      ORS_PROFILE VARCHAR(30), SHIFT_TYPE VARCHAR(30),
+      SHIFT_START_HOUR INT, SHIFT_END_HOUR INT,
+      HOME_LOCATION_ID VARCHAR, DRIVER_PROFILE VARCHAR(20),
+      OPERATING_MODE VARCHAR(30), BASE_SPEED_KMH FLOAT, BATTERY_RANGE_KM FLOAT,
+      JOB_ID VARCHAR
+    ) COMMENT = '{"origin":"sf_sit-is-fleet","name":"oss-build-routing-solution","version":{"major":1,"minor":0},"attributes":{"is_quickstart":1,"source":"sql"}}'`, db: UNIFIED_DB, schema: UNIFIED_SCHEMA },
+    { sql: `CREATE TABLE IF NOT EXISTS ${UNIFIED_DB}.${UNIFIED_SCHEMA}.DIM_POIS (
+      LOCATION_ID VARCHAR, REGION VARCHAR(100), NAME VARCHAR,
+      LOCATION_TYPE VARCHAR(30), CATEGORY VARCHAR(50),
+      LAT FLOAT, LNG FLOAT, POINT_GEOM GEOGRAPHY, SOURCE VARCHAR(20),
+      JOB_ID VARCHAR
+    ) COMMENT = '{"origin":"sf_sit-is-fleet","name":"oss-build-routing-solution","version":{"major":1,"minor":0},"attributes":{"is_quickstart":1,"source":"sql"}}'`, db: UNIFIED_DB, schema: UNIFIED_SCHEMA },
+    { sql: `CREATE TABLE IF NOT EXISTS ${UNIFIED_DB}.${UNIFIED_SCHEMA}.DIM_TRIP_SCHEDULE (
+      SCHEDULE_ID VARCHAR, VEHICLE_ID VARCHAR, DRIVER_ID VARCHAR,
+      VEHICLE_TYPE VARCHAR(20), REGION VARCHAR(100),
+      TRIP_DATE DATE, TRIP_SEQ INT,
+      ORIGIN_POI_ID VARCHAR, DESTINATION_POI_ID VARCHAR,
+      PLANNED_START TIMESTAMP_NTZ, PLANNED_END TIMESTAMP_NTZ,
+      SHIFT_TYPE VARCHAR(30), ORS_PROFILE VARCHAR(30),
+      DISTANCE_KM FLOAT, DURATION_MINUTES FLOAT, STATUS VARCHAR(20),
+      JOB_ID VARCHAR
+    ) COMMENT = '{"origin":"sf_sit-is-fleet","name":"oss-build-routing-solution","version":{"major":1,"minor":0},"attributes":{"is_quickstart":1,"source":"sql"}}'`, db: UNIFIED_DB, schema: UNIFIED_SCHEMA },
+    { sql: `CREATE TABLE IF NOT EXISTS ${UNIFIED_DB}.${UNIFIED_SCHEMA}.FACT_FREIGHT_OFFERS (
+      OFFER_ID VARCHAR, REGION VARCHAR(100), VEHICLE_TYPE VARCHAR(20),
+      SOURCE VARCHAR(30),
+      PICKUP_POI_ID VARCHAR, PICKUP_LAT FLOAT, PICKUP_LON FLOAT, PICKUP_GEOM GEOGRAPHY,
+      DROPOFF_POI_ID VARCHAR, DROPOFF_LAT FLOAT, DROPOFF_LON FLOAT, DROPOFF_GEOM GEOGRAPHY,
+      PICKUP_FROM_TS TIMESTAMP_NTZ, PICKUP_TO_TS TIMESTAMP_NTZ,
+      WEIGHT_KG NUMBER, PRODUCT VARCHAR, PRICE_USD NUMBER, HAZMAT BOOLEAN,
+      LISTING_TEXT VARCHAR, POSTED_AT TIMESTAMP_NTZ,
+      JOB_ID VARCHAR
+    ) COMMENT = '{"origin":"sf_sit-is-fleet","name":"oss-build-routing-solution","version":{"major":1,"minor":0},"attributes":{"is_quickstart":1,"source":"sql"}}'`, db: UNIFIED_DB, schema: UNIFIED_SCHEMA },
+    { sql: `CREATE TABLE IF NOT EXISTS FLEET_INTELLIGENCE.CORE.GENERATION_JOBS (
+      JOB_ID VARCHAR, PRESET_NAME VARCHAR, REGION VARCHAR(100),
+      ORS_PROFILE VARCHAR(30), NUM_VEHICLES INT,
+      START_DATE VARCHAR, END_DATE VARCHAR,
+      STATUS VARCHAR(20), CONFIG VARIANT,
+      POINTS_GENERATED INT DEFAULT 0, TRIPS_GENERATED INT DEFAULT 0,
+      ERROR_MESSAGE VARCHAR, STARTED_AT TIMESTAMP_NTZ DEFAULT SYSDATE(),
+      COMPLETED_AT TIMESTAMP_NTZ, LOG_TEXT VARIANT
+    ) COMMENT = '{"origin":"sf_sit-is-fleet","name":"oss-build-routing-solution","version":{"major":1,"minor":0},"attributes":{"is_quickstart":1,"source":"sql"}}'`, db: 'FLEET_INTELLIGENCE', schema: 'CORE' },
+    { sql: `ALTER TABLE FLEET_INTELLIGENCE.CORE.GENERATION_JOBS ADD COLUMN IF NOT EXISTS LOG_TEXT VARIANT`, db: 'FLEET_INTELLIGENCE', schema: 'CORE' },
+  ];
+  for (const { sql, db, schema } of ddls) {
+    try {
+      await snowSql(sql, db, schema);
+    } catch (e: any) {
+      const raw = e.message || '';
+      if (raw.includes('Insufficient privileges') || raw.includes('42501') || raw.includes('access control')) {
+        const hint = `Missing privileges on ${db}.${schema}. ` +
+          `Run the Data Studio setup SQL from SKILL.md Step 6.3 as ACCOUNTADMIN, ` +
+          `or re-run deploy.sh which grants all required privileges automatically.`;
+        log('ERROR', 'Studio', hint);
+        throw new Error(hint);
+      }
+      const msg = `DDL error (${db}.${schema}): ${raw.slice(0, 200)}`;
+      console.error(`[Studio] ${msg}`);
+      log('ERROR', 'Studio', msg);
+      throw new Error(msg);
+    }
+  }
+}
